RandImpl insert() now properly merges if exists.
Although it's a fake implementation, the ELFT API indicates that
insert() should merge if a template with the same identifier already
exists in the reference database. This is to show an example of how
an implementation might refactor ExtractionInterface::mergeTemplates()
into an area that can be used from SearchInterface for use in this
situation.