Explicitly set --enable-new-dtags linker flag.
Evaluation code running on NIST systems is compiled with mpicxx, which
sets -Wl,--enable-new-dtags in LDFLAGS. If a participant is looking in
RPATH for associated libraries, they would find them in validation but
not on NIST systems. Conversely, if they looked in RUNPATH, they would
find them on NIST systems but not during validation. This change
should create consistency between the two.